+++ public/flux/store.js
@@ -1,28 +1,15 @@
-import {SYSTYPES} from './types.js'
+import createStore from "./create_store.js";
+import {authorizeReducer, changePageReducer, mainPageReducer, combineReducers} from './reducers.js';
 
-export default function createStore(rootReducer, initialState) {
-  if (fluxDebug) {
-    console.log('store creation | rootReducer: ', rootReducer, 'initialState: ', initialState)
-  }
-  let state = rootReducer(initialState, {type: SYSTYPES.INIT})
-  if (fluxDebug) {
-    console.log('store creation | initialState after INIT: ', initialState)
-  }
-  const subscribers = []
-
-  return {
-    dispatch(action) {
-      state = rootReducer(state, action)
-      subscribers.forEach(subscriber => subscriber())
-      if (fluxDebug) {
-        console.log(`state after dispatch: ${JSON.stringify(state)}`)
-      }
-    },
-    subscribe(callback) {
-      subscribers.push(callback)
-    },
-    getState() {
-      return state
-    }
-  }
+class Store {
 }
+
+const store = new Store();
+Object.assign(store,
+  createStore(combineReducers({
+    authorization: authorizeReducer,
+    page: changePageReducer,
+    mainPage: mainPageReducer,
+  })));
+
+export default store;
